refactor repository helper
